Bug fixes, quality of life
Min SDK Version changed by 1 version older (so I can test on the note 8)
Viewing portion of the page changed to be exact size of screen to fill the area, and adjusted the position of the pressures.
Changed the names of the quick select icons to a more functional name, even though they are hardcoded and it's not super ideal... we are running on low arduino space!
Fixed the bug where the pressure scroll wheels would be blank sometimes.
Fixed the part of the code that makes it retry connections so it will properly retry. Seems to function well now?
Tried to add code to early-end the connection to make it retry quicker but it didn't work, caused issues of course.