Added a redundancy check in the solenoid code
Wanted to add redundancy so I can call it multiple times without actually calling the digitalWrite function if it doesn't need to be called... I think this should be fine.
Worth mentioning that I added this while thinking of how the compressor code is written, check the part where it closes in the compressor code to see how they are related. It will continuously call Solenoid::close() when the max pressure is reached.