Simplified bluetooth code
Some of this code doesn't appear to be necessary anymore. Maybe it was for getting rid of bugs on the arduino nano? im not too sure. Seems to work fine with these updates.
Also added a buffer length check and a delay in case it gets stuck waiting forever for a new line, we don't want it in an infinite loop without a delay.