fixed wii fit code
accidentally had up instead of down for the down code.
also called balanceBoardCode on the opposite sode of what's being pressed in order to reset it
also set boolean to static so it doesn't automatically reset the right side i guess? idk, probably not necessary.
Based on debug output the code appears to work fine now.

in the future we can change it so we can press both sides at the same time. for now this is fine.